+++ tests/testthat/test-clustering.R
@@ -144,6 +144,16 @@ test_that("kNN clustering works correctly", {
   expect_true(length(knn) > 0 )
 })
 
+
+test_that("kMeans clustering - no distance scores", {
+  expect_true(is.null(kMeans_clustering(NULL, 3)))
+})
+
+test_that("kMeans clustering works correctly", {
+  kMeans <- kMeans_clustering(distances, 1)
+  expect_true(length(kMeans) > 0  & length(kMeans) <= 1)
+})
+
 test_that("getClusterDataTable - no geneset names", {
   expect_error(.getClusterDatatable(list(), list()))
 })